Moved userExists tests to SecurityTest File
