+++ Smart-Contract-Verifier-Server/src/main/resources/application.properties
@@ -1,8 +1,8 @@
-spring.datasource.url=jdbc:postgresql://${dburl}/verifierserver
-spring.datasource.username=${dbuser}
-spring.datasource.password=${dbpassword}
+spring.datasource.url=jdbc:postgresql://localhost:5432/verifierserver
+spring.datasource.username=postgres
+spring.datasource.password=smartcontractserverdev
 spring.datasource.driver-class-name=org.postgresql.Driver
-spring.sql.init.enabled=${dbgenschema}
+spring.sql.init.enabled=false
 spring.sql.init.schema-locations=classpath:schema.sql
 
 server.port=8080
